improved computation of weak splitting field
The computation of the weak splitting field in fake_padic_completion
has been improved a lot: we do not compute a complete list of
approximate irreducible factors; instead we do a MacLane approximation
and then choose one nice factor, which is then approximated as far
as necessary.

Also fixed a few bugs in fake_padic_completion.