Removed an assert command
In WeakpAdicGaloisExtension.__init__ we have removed the test
that the degree of the ramified extension added at the end
to achieve the 'minimal_ramification' has to be prime to p.
There are situations where it is necessary to have this degree
not prime to p. The problem may be that the extension obtained
at the end may not be weak Galois.

We have to reconsider this at some point.