Very provisional fix of a problem in superp.
The problem occurs when computing the etale locus and one of the
functions delta[i] are constant. The rational domain given by
vK(delta[i])>=0 is then either the whole Berkovich line, or empty.
This would make sense here, but calling RationalDomainOnBerkovichLine
leads to an error.

I added an if statement to catch this exception, but what really should
be fixed is RationalDomainOnBerkovichLine, which should allow a
defining function which is constant.