mock modules without tricking sphinx
which refuses to build subclasses of magic mocks for some reason. The reason is
not the omission of inherited methods. I guess that it tries to find something
on these classes and then decides not to render them because it (naturally)
find whatever it's looking for on a mock.
Note that sphinx's built-in mocking has the same issue.