Started major rewriting of `berkovich_lines`
The main goal is to fix issue #39: allow points of type I
whose Galois conjugates are both inside and outside the
unit disk. Such a point is given by a pseudo-valuation v
on the function field F=K(x) which is induced from a pseudo-
valuation v0 on a subring K[y], where y is an arbitrary parameter
of F (typically y=x/c or y = 1/x).

One can now call the classes `TypeIPointOnBerkovichLine`
and `TypeIIpointOnBerkovichLine` also by passing instead of v
a pair (v0, y).

Points are more systematically represented by a discoid D, given
by a pair (f, s), where f is either a polynomial in x, or f=1/x, and
such that D is either contained in or disjoint from the unit disk.
Then D is uniquely associated to the point, and this representation
is used to check for inequalities and compute infima.

There is still a lot to do, in particular

- inequalities and infima for type-I-points
- more docstrings
- remove obsolete code
- test!