Experimental code to fix issue #40
I have added to ``reduction_tree`` a new function
`mac_lane_approximants_experimental(v,G)` which replaces
`v.mac_lane_approximants(G)` for the computation of upper components.

The function does the usual `mac_lane_steps`, but after each step
simplifies the resulting indictive valuations, using a function
`simplify_inductive_valuation(v)`.

The latter function just replaces the last key polynomial `phi` of an
inductive polynomial by a (heuristically) simpler polynomial `phi1`
which is close enough to phi to not change the valuation.

The simplification of `phi` is done via a function
`simplify_wrt_valuation(v, a, N)` which finds an element `b` of the
domain of `v` (a ring which also contains `a`) such that
`v(a-b) > N` and such that `b` is heuristically simpler than `a`.