fix issue 61
In `reduction_trees.py`, `LowerComponent.upper_components`,
the call of `v.mac_lane_approximants` now has the parameter
`require_incomparability=True`. This seems to fix the problem.

However, one still needs to make sure that this parameter is
enough to guarantee that the approximants are sufficiently
precise to induce the desired function field valuation.