improve computation of field of constant degree
Given a smooth projective curve Y with function field F,
we have a more reliable method to compute the degree of
the field of constants, as an extension of the constant
base field.
If the constant base field is finite, we test for all
possible finite extensions whether the defining equation
is irreducible or not. Actually, this algorithm gives the
field of constants itself.
For a number field, we reduce the equation at 10 places
of good reduction and apply the method for finite fields.

ToDo:
- documentation and meaningful doctests
- NotImplementedError for fields which are
  neither finite nor number fields