Improve computation of the genus of a curve.
Given a smooth projective curve over a field of characteristic
zero, we can use the 'tame' Riemann Hurwitz Formula to compute
the genus. For this, we only need to know the ramification
and inertia degrees of the valuations corresponding to the
ramification points. For a curve over a number field, this is
sometimes must faster.