next try
there is now a new version of 
`make_finite_field` and `make_function_field`,
which is a hack to solve issue #103. Adaptions
to this in several places. 

It may be an option to keep these two functions
in an experimental version until https://trac.sagemath.org/ticket/26103 
is completed.
However, the whole module `reduction_trees` needs to 
be reviewed first.