thorough rewriting of `reduction_trees`
In this new version of reduction_trees, 
the maps between different components
(inertial, lower and upper) are constructed
in a systematic way and such that they are 
compatible with the maps of generic fibers.

All components are created as curves over an
*absolute* finite fields, thus fixing issue #103
and adressing in aprt issue #66. This is done
by using the helper functions and  
`make_finite_fields` which turns a finite field into an absolute fielte, 
and `make_function_fields`
which turns a function field into a 'true' function
field over an absolute finite field. 

There should be more doctests and internal 
consistency test.