corrected bug in morphism_of_smooth_projective_curves
In the previous version, a nontrivial map between two
projective lines was mistakenly regarded as the identity
because the two function fields were regarded as the same
by Sage. This has been fixed.

In this process, the definition of and the distinction between the two 
cases where a morphism is defined has 
been made clearer and more explicit. The documentation 
was expanded. 

The class ``SmoothProjectiveCurve`` has a new function
``structure_map`` which returns the canonical map from
the curve to the projective line.