Adapt mclf to newer Sage version
The commit tries to adapt mclf to several changes that have occured with 
the lates versions of Sage. 

- it seems that residue fields of valuations on function fields are now 
realized as "function fields" if this makes sense. Therefore, the 
function `make_function_field' in `reduction_trees' is now probably 
superfluous. I haven't gotten rid of it, but just made it give back the 
identical field if the input is already a function field. 
- adapted some doctest
- minor edits, to please flakes8