Fix bug related to issue #103
With Sage 8.6, residue fields of valuations on function fields are now 
function fields themselve, if this makes sense. However, the constant 
base field may not be a *true* finite field (this is trac:26103). This 
resulted in an error message when applying the method 
``make_function_field``.

Now ``make_function_field`` creates a new function field with the right 
kind of constant base field if the input is already a function field.