New module `valuative_functions`
In this module, a class `ValuativeFunctions` is defined. The goal is to 
simplify and speed up the computation of rational domains, in particular 
in `superp_models`.

The basic infrastructure and the initialization process for this class 
are defined, but not tested. Nothng works at the moment. One thing that 
is missing is the method `is_equal` for points of type V.