new module `piecewise_affine_functions`
This is a new and improved version of `valuative_function`. The main new 
class is
`PiecewiseAffineFunction`, and the valuative functions are objects in 
this class which can be created with `valuatve_function(D, f)`. Other 
new things:
- auxiliary classes `Domain`, `Discoid`, `DirectedPath` and 
`AffineFunction`

Changes in `berkovich_line`:
- correction in `point_from_discoid'
- correction in 'is_incomparable'
- new option `require_maximal_degree` in `approximation`

The creation and evaluation of piecewise affine functions works quite 
well. Other features are untested.