Major rewriting of `affinoid_domains`
The main goal was to speed up the computation of rational domains and of 
the union of affinoid domains. The second goal is more or less reached, 
by using the class `UnionOfDomains`.

This is the first commit where things work. There will be likely many 
more changes and improvements. 

To Do:
- more doctests
- more comments
- class `ValuativeDomain` to improve computation of rational domains
- fix bug in `point_close_to_boundary`