Compute genus via reduction
The method `genus` accepts a flag `use_reduction`. if this is `True` and the curve is defined over a number field, the genus is computed by reduction to a good prime.

Also, if the curve is defined over a finite field, but the characteristic is larger then the covering degree, then the tame RH formula is used, just as in characteristic zero.