New implementation of Factoring polynomials over function fields
Since the old factoring implementation sometimes gives an error due to a bug in Sage, we implemented a new one.
It is based on an experimental implementation in Sage by Julian Rüth.
Performance is very bad, because gcd computaton of polynomials over function fields is not well implmented in Sage.