Some progress:
we implemented the computation of the xgcd of polynomials over a UFD; using either the primitive PRS or the subresultant PRs algorithm.

If the UFD is QQ or a function field over a finite field, the performance is very reasonable. 

Much better performance for large examples could be obtained by a modular approach.
But it is probably not worth the effort to do this in general. But it should be easy to implement a very fast test whther two polynomials are relatively prime, using reduction to some random "good" place of the base field.

The most urgent task is to extend the computation of the gcd (not necessarily the xgcd) to polynomials over nonrational function fields.